refactor(tests): centralize mock_project_root fixture in conftest.py
- Add shared mock_project_root fixture in tests/conftest.py
- Add reset_project_root_cache fixture to clear cache between tests
- Remove duplicate mock_project_root fixtures from:
  - tests/configuration/test_experiment_factory.py
  - tests/configuration/test_configuration.py
  - tests/configuration/test_experiment_group.py
